fix: fix CreateTransactionController required fields validation
