+++ src/errors/user.js
@@ -7,7 +7,7 @@ export class EmailAlreadyInUseError extends Error {
 }
 export class UserNotFoundError extends Error {
   constructor(userId) {
-    super(`User not found with ID: ${userId}`)
+    super(`User with id: ${userId} not found`)
     this.name = 'UserNotFoundError'
   }
 }
