+++ src/repositories/postgres/transaction/create-transaction.js
@@ -1,8 +1,8 @@
-import PostgresHelper from '../../helpers/postgres-helper.js'
+import { PostgresHelper } from '../../../db/postgres/helper.js'
 export class PostgresCreateTransactionRepository {
   async execute(createTransactionParams) {
     const createdTransaction = await PostgresHelper.query(
-      `INSERT INTO transaction (id, user_id, name, date, amount, type) 
+      `INSERT INTO transactions (id, user_id, name, date, amount, type) 
       VALUES ($1, $2, $3, $4, $5, $6) RETURNING *`,
       [
         createTransactionParams.id,
@@ -13,6 +13,6 @@ export class PostgresCreateTransactionRepository {
         createTransactionParams.type,
       ],
     )
-    return createdTransaction[0] || null
+    return createdTransaction[0]
   }
 }
