+++ src/use-cases/index.js
@@ -3,3 +3,5 @@ export * from './user/create-user.js'
 export * from './user/get-user-by-id.js'
 export * from './user/update-user.js'
 export * from './user/delete-user.js'
+
+export * from './transaction/create-transaction.js'
